Improve coding standard and refactor parse() method
Currently, the code has poor coding standard.

Improving the coding standard will make it easier to find bugs.

Let's:
- improve the coding standard as instructed.
- fix the bugs that become apparent during refactoring.
