+++ src/main/java/bot/utils/Storage.java
@@ -1,13 +1,12 @@
 package bot.utils;
 
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.*;
 import java.util.ArrayList;
 import java.util.Scanner;
 
 import bot.exceptions.InvalidTaskException;
 import bot.exceptions.LoadingException;
+import bot.utils.tasks.Task;
 
 /**
  * Abstraction for storing and loading data from a local file.
@@ -16,7 +15,7 @@ public class Storage {
     /**
      * File path for loading and saving data.
      */
-    private String filePath;
+    private final String filePath;
 
     /**
      * Creates a Storage object that stores and loads data at that file path.
@@ -35,28 +34,50 @@ public class Storage {
      * @throws LoadingException If the data file cannot be read or created.
      */
     public ArrayList<Task> load(Ui ui) throws LoadingException {
-        ArrayList<Task> tasks = new ArrayList<>();
         File f = new File(this.filePath);
-        File parent = new File(f.getParent());
+        ArrayList<Task> tasks;
         try {
-            if (f.isFile()) {
-                Scanner scanner = new Scanner(f);
-                while (scanner.hasNextLine()) {
-                    try {
-                        tasks.add(Task.convertFromString(scanner.nextLine()));
-                    } catch (InvalidTaskException e) {
-                        ui.println("Invalid task, skipping...");
-                    }
-                }
-            } else {
-                ui.println("No data found, creating...");
-                if (!parent.isDirectory()) {
-                    parent.mkdirs();
-                }
-                f.createNewFile();
+            if (!f.isFile()) {
+                createTasksFile(f);
             }
+            tasks = loadTasksFromFile(f, ui);
         } catch (IOException | SecurityException e) {
             throw new LoadingException(e.getMessage());
+        }
+        return tasks;
+    }
+
+    /**
+     * Creates the task file in the specified directory given by the file's filepath.
+     *
+     * @param f Details of file to create.
+     * @throws IOException If the file cannot be created.
+     */
+    private static void createTasksFile(File f) throws IOException {
+        File parent = new File(f.getParent());
+        if (!parent.isDirectory()) {
+            parent.mkdirs();
+        }
+        f.createNewFile();
+    }
+
+    /**
+     * Creates an ArrayList of tasks from the contents of the file, printing to the Ui object given.
+     *
+     * @param f File to read tasks from.
+     * @param ui UI to print to.
+     * @return Tasks in an ArrayList.
+     * @throws FileNotFoundException If file cannot be found.
+     */
+    private static ArrayList<Task> loadTasksFromFile(File f, Ui ui) throws FileNotFoundException {
+        Scanner scanner = new Scanner(f);
+        ArrayList<Task> tasks = new ArrayList<>();
+        while (scanner.hasNextLine()) {
+            try {
+                tasks.add(Task.convertFromDataString(scanner.nextLine()));
+            } catch (InvalidTaskException e) {
+                System.out.println("Invalid task, skipping...");
+            }
         }
         return tasks;
     }
@@ -71,6 +92,7 @@ public class Storage {
         try {
             File f = new File(this.filePath);
             FileWriter fw = new FileWriter(f);
+            // We did it like this so the last line isn't a line separator.
             for (int i = 1; i < lst.size(); i++) {
                 fw.write(lst.get(i).convertToDataString());
                 fw.write(System.lineSeparator());
