+++ src/test/java/bot/utils/TaskListTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 
 import java.util.ArrayList;
 
+import bot.utils.tasks.Task;
 import org.junit.jupiter.api.Test;
 
 import bot.exceptions.InvalidIndexException;
