--- text-ui-test/input.txt
@@ -1,20 +1,0 @@
-hello
-list
-delete 1
-delete 1
-todo taskone
-list
-mark 1
-list
-unmark 1
-list
-deadline tasktwo /by 2023-11-11
-list
-event taskthree /from 2023-08-25 /to 2023-08-26
-event taskfail /from 2020-11-22 /to 2019-11-11
-deadline taskfail /by 2020-20-20
-list
-mark 3
-delete 3
-list
-bye
